Move countries.json update into separate job
This allows it to be run in paralel without so failures in this job
don't affect other jobs and vice versa.
