+++ app/jobs.rb
@@ -3,3 +3,4 @@ require 'app/jobs/accept_submission_job'
 require 'app/jobs/deploy_viewer_sinatra_pull_request_job'
 require 'app/jobs/handle_everypolitician_data_pull_request_job'
 require 'app/jobs/send_webhook_job'
+require 'app/jobs/update_countries_json_job'
