+++ app/jobs/handle_everypolitician_data_pull_request_job.rb
@@ -1,7 +1,6 @@
 # Checks for everypolitician-data pull requests and creates deployment events
 class HandleEverypoliticianDataPullRequestJob
   include Sidekiq::Worker
-  include Github
 
   attr_accessor :pull_request
   attr_reader :github
@@ -51,13 +50,7 @@ class HandleEverypoliticianDataPullRequestJob
 
   def update_countries_json
     branch = pull_request['pull_request']['head']['ref']
-    message = 'Refresh countries.json'
-    with_git_repo(everypolitician_data_repo, branch: branch, message: message) do
-      # Unset bundler environment variables so it uses the correct Gemfile etc.
-      env = {'BUNDLE_GEMFILE' => nil, 'BUNDLE_BIN_PATH' => nil, 'RUBYOPT' => nil, 'RUBYLIB' => nil}
-      system(env, 'bundle install')
-      system(env, 'bundle exec rake countries.json')
-    end
+    UpdateCountriesJsonJob.perform_async(branch)
   end
 
   def merged?
@@ -73,13 +66,5 @@ class HandleEverypoliticianDataPullRequestJob
 
   def everypolitician_data_repo
     ENV['EVERYPOLITICIAN_DATA_REPO']
-  end
-
-  class FailedSystemCall < StandardError; end
-
-  def system(*args)
-    if !Kernel.system(*args)
-      raise FailedSystemCall, "#{args} exited with #$?"
-    end
   end
 end
